[FIX] linting and style issues base_delivery_carrier_files
