+++ base_delivery_carrier_files/csv_writer/__init__.py
@@ -1,1 +1,3 @@
+# -*- coding: utf-8 -*-
+
 from .unicode_writer import UnicodeWriter
