+++ base_delivery_carrier_files/csv_writer/unicode_writer.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # CSV writer from python doc
 import csv
 import codecs
