+++ base_delivery_carrier_files/models/__init__.py
@@ -2,7 +2,7 @@
 # Copyright 2012 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import csv_writer
-from . import generator
-from . import models
-from . import wizards
+from . import delivery_carrier
+from . import delivery_carrier_file
+from . import stock_move
+from . import stock_picking
