+++ base_delivery_carrier_files/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Guewen Baconnier <guewen.baconnier@camptocamp.com>
+* Tom Blauwendraat <tom@sunflowerweb.nl>
