+++ base_delivery_carrier_files/views/stock_view.xml
