+++ base_delivery_carrier_files/wizards/__init__.py
@@ -2,7 +2,4 @@
 # Copyright 2012 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import csv_writer
-from . import generator
-from . import models
-from . import wizards
+from . import generate_carrier_files
